fix "Waiting for network..." bug
this change forces steam to run in the old UI because the new UI never loads for some people inside of distrobox. This is only an issue when you start steam for the first time to install steamVR. the new UI is able to be loaded after this though, so it's only necessary here.